requirements6.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Django==5.1.4
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.5
langsmith==0.2.10
MarkupSafe==3.0.2
marshmallow==3.25.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy>=1.25.0
openai==1.59.6
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
sympy==1.13.1
tenacity==9.0.0
toposort==1.10
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
async-timeout>=4.0.0,<5.0.0 
